initializint of example 2
