+++ recognizer/__init__.py
@@ -1,2 +1,2 @@
 from .offline import * # noqa
-from .online import * # noqa
+# from .online import * # noqa
