+++ words.py
@@ -1,5 +1,5 @@
 
-TRIGGERS = {"джарвис", "djarvis"}
+TRIGGERS = {"джарвис", "djarvis", "чарльз"}
 
 # Теперь значения — это имена функций-обработчиков (строки)
 data_set = {
